requirements.txt
certifi==2019.6.16
chardet==3.0.4
ddt==1.2.1
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
mysql-connector-python==8.0.17
openpyxl==2.6.2
protobuf==3.9.0
PyMySQL==0.9.3
requests==2.22.0
six==1.12.0
urllib3==1.25.3
